Update Action to NodeJS 20 (#2)
According to [GitHub Actions: Transitioning from Node 16 to Node 20 - The GitHub Blog](https://github.blog/changelog/2023-09-22-github-actions-transitioning-from-node-16-to-node-20/), GitHub is deprecating Node16 from their runners.

Today, use of this action shows the following warning: `Node.js 16 actions are deprecated. Please update the following actions to use Node.js 20: Mojang/auto-request-review@v1`.

Updates the action to use NodeJS 20. Validated locally that Node 20 can run the UTs, and updated the CI to run under Node 20.

This also fixes a misc issue with linting flagging the `??=` operator that was not caught in #1 due to the CI not running.